+++ src/ActionGameCore/Assets/CharacterControls/Scripts/Runtime/Inputs/IInputReceiver.cs
@@ -1,0 +1,7 @@
+namespace CharacterControls.Inputs
+{
+    public interface IInputReceiver
+    {
+        void OnReceiveInput(InputContext context);
+    }
+}