+++ src/ActionGameCore/Assets/CharacterControls/Scripts/Runtime/Movements/Modules/CharacterModuleBase.cs
@@ -1,20 +1,37 @@
+using CharacterControls.Inputs;
 using UnityEngine;
 
 namespace CharacterControls.Movements.Modules
 {
-    public abstract class CharacterModuleBase : MonoBehaviour, ICharacterModule
+    public abstract class CharacterModuleBase : MonoBehaviour, ICharacterModule, IInputReceiver
     {
-        protected virtual void Start()
+        private CharacterMoveInput _characterInput;
+
+        protected virtual void OnEnable()
         {
+            _characterInput = GetComponentInParent<CharacterMoveInput>();
+            if (_characterInput != null)
+            {
+                _characterInput.RegisterReceiver(this);
+            }
+
             var characterMoveController = GetComponentInParent<CharacterMoveController>();
             characterMoveController?.RegisterModule(this);
         }
 
-        protected virtual void OnDestroy()
+        protected virtual void OnDisable()
         {
+            if (_characterInput != null)
+            {
+                _characterInput.UnregisterReceiver(this);
+                _characterInput = null;
+            }
+
             var characterMoveController = GetComponentInParent<CharacterMoveController>();
             characterMoveController?.UnregisterModule(this);
         }
+
+        public abstract void OnReceiveInput(InputContext context);
 
         public abstract void FixedUpdateModule(in CharacterMoveModulePayload payload);
     }